refactor: adjust config/persistence
