+++ src/core/4.main/config/persistence/databases/index.ts
@@ -1,6 +1,6 @@
 import { Persistence } from '@/core/4.main/config/config'
-import { mongodbPersistence } from '@/core/4.main/config/persistence/mongodb-persistence'
-import { postgresPersistence } from '@/core/4.main/config/persistence/postgres-persistence'
+import { mongodbPersistence } from '@/core/4.main/config/persistence/databases/mongodb-persistence'
+import { postgresPersistence } from '@/core/4.main/config/persistence/databases/postgres-persistence'
 
 const DATABASE = process.env.DATABASE
