+++ src/core/4.main/config/persistence/factories/mongodb-factory.ts
@@ -1,4 +1,4 @@
-import { DatabaseFactories } from '@/core/4.main/config/database-factories'
+import { DatabaseFactories } from '@/core/4.main/config/persistence/factories'
 import { MongodbUserFactory } from '@/user/3.infra/persistence/mongodb/factories/mongodb-user-factory'
 
 export const mongodbFactories: DatabaseFactories = {
