+++ src/core/4.main/config/persistence/factories/postgres-factory.ts
@@ -1,4 +1,4 @@
-import { DatabaseFactories } from '@/core/4.main/config/database-factories'
+import { DatabaseFactories } from '@/core/4.main/config/persistence/factories'
 import { PostgresUserFactory } from '@/user/3.infra/persistence/postgres/factories/postgres-user-factory'
 
 export const postgresFactories: DatabaseFactories = {
