+++ test/modules/user/3.infra/api/routes/change-password.integration.test.ts
@@ -4,7 +4,7 @@ import { TokenType } from '@/core/1.application/cryptography/encrypter'
 import { Route, WebApp } from '@/core/3.infra/api/app/web-app'
 import { DatabaseFactory } from '@/core/3.infra/persistence/database-factory'
 import { config } from '@/core/4.main/config/config'
-import { factories } from '@/core/4.main/config/database-factories'
+import { factories } from '@/core/4.main/config/persistence/factories'
 import { authMiddlewareFactory } from '@/core/4.main/factories/auth-middle-factory'
 import { schemaValidatorMiddlewareFactory } from '@/core/4.main/factories/schema-validator-middleware-factory'
 import { UserAggregateCreateParams } from '@/user/0.domain/aggregates/user-aggregate'
