+++ test/modules/user/3.infra/api/routes/sign-in.integration.test.ts
@@ -3,7 +3,7 @@ import request from 'supertest'
 import { Route, WebApp } from '@/core/3.infra/api/app/web-app'
 import { DatabaseFactory } from '@/core/3.infra/persistence/database-factory'
 import { config } from '@/core/4.main/config/config'
-import { factories } from '@/core/4.main/config/database-factories'
+import { factories } from '@/core/4.main/config/persistence/factories'
 import { schemaValidatorMiddlewareFactory } from '@/core/4.main/factories/schema-validator-middleware-factory'
 import { UserAggregateCreateParams } from '@/user/0.domain/aggregates/user-aggregate'
 import { signInRoute } from '@/user/3.infra/api/routes/sign-in-route'
