+++ test/modules/user/3.infra/persistence/postgres/repositories/postgres-user-repository.integration.test.ts
@@ -1,7 +1,7 @@
 import { DatabaseFactory } from '@/core/3.infra/persistence/database-factory'
 import { postgres } from '@/core/3.infra/persistence/postgres/client/postgres-client'
-import { postgresFactories } from '@/core/4.main/config/database-factories/postgres-factory'
-import { postgresPersistence } from '@/core/4.main/config/persistence/postgres-persistence'
+import { postgresPersistence } from '@/core/4.main/config/persistence/databases/postgres-persistence'
+import { postgresFactories } from '@/core/4.main/config/persistence/factories/postgres-factory'
 import { UserAggregate, UserAggregateCreateParams } from '@/user/0.domain/aggregates/user-aggregate'
 import { EmailConfirmed } from '@/user/0.domain/value-objects/email-confirmed'
 import { PostgresUserRepository } from '@/user/3.infra/persistence/postgres/repositories/postgres-user-repository'
